Fixing incorrect placement to CPU when offload_to_disk=False during initial input capture (looks like a bug, but provide lower VRAM usage), so thi is not final.
observed when offload_to_disk=False - the input capture proceed on CPU device and inputs are also on CPU (probably by mistake). While with offload_to_disk=True - it proceed on cuda device.
Idk what is better, the VRAM usage is lower when it is on CPU - during inputs capture and during quantization (at least for layer 0), but need to compare speed of forward pass/input capture.
Looks like a bug to me.
As we have implemented a feature of calibration_data_device in separate branch the placement will be rearranged anyway.